+++ views/layouts/main.php
@@ -9,6 +9,7 @@ use yii\bootstrap4\Html;
 use yii\bootstrap4\Nav;
 use yii\bootstrap4\NavBar;
 use app\assets\AppAsset;
+use yii\helpers\Url;
 
 AppAsset::register($this);
 
@@ -22,7 +23,6 @@ AppAsset::register($this);
     <meta name="viewport" content="width=device-width,initial-scale=1">
     <?php $this->registerCsrfMetaTags() ?>
     <title><?= Html::encode($this->title) ?></title>
-    <!-- <link rel="stylesheet" href="css/style.css"> -->
     <?php $this->head() ?>
 </head>
 
@@ -34,45 +34,50 @@ AppAsset::register($this);
             <a href='#' class="header-logo">
                 <img class="logo-image" src="/img/logotype.png" width=227 height=60 alt="taskforce">
             </a>
-            <div class="nav-wrapper">
-                <ul class="nav-list">
-                    <li class="list-item list-item--active">
-                        <a class="link link--nav">Новое</a>
-                    </li>
-                    <li class="list-item">
-                        <a href="#" class="link link--nav">Мои задания</a>
-                    </li>
-                    <li class="list-item">
-                        <a href="#" class="link link--nav">Создать задание</a>
-                    </li>
-                    <li class="list-item">
-                        <a href="#" class="link link--nav">Настройки</a>
-                    </li>
-                </ul>
-            </div>
-        </nav>
-        <div class="user-block">
-            <a href="#">
-                <img class="user-photo" src="/img/man-glasses.png" width="55" height="55" alt="Аватар">
-            </a>
-            <div class="user-menu">
-                <p class="user-name">Василий</p>
-                <div class="popup-head">
-                    <ul class="popup-menu">
-                        <li class="menu-item">
-                            <a href="#" class="link">Настройки</a>
+            <?php if (Url::current() !== Url::to(['site/registration'])) : ?>
+                <div class="nav-wrapper">
+                    <ul class="nav-list">
+                        <li class="list-item list-item--active">
+                            <a class="link link--nav">Новое</a>
                         </li>
-                        <li class="menu-item">
-                            <a href="#" class="link">Связаться с нами</a>
+                        <li class="list-item">
+                            <a href="#" class="link link--nav">Мои задания</a>
                         </li>
-                        <li class="menu-item">
-                            <a href="#" class="link">Выход из системы</a>
+                        <li class="list-item">
+                            <a href="#" class="link link--nav">Создать задание</a>
                         </li>
-
+                        <li class="list-item">
+                            <a href="#" class="link link--nav">Настройки</a>
+                        </li>
                     </ul>
                 </div>
+            <?php endif; ?>
+        </nav>
+
+        <?php if (Url::current() !== Url::to(['site/registration'])) : ?>
+            <div class="user-block">
+                <a href="#">
+                    <img class="user-photo" src="/img/man-glasses.png" width="55" height="55" alt="Аватар">
+                </a>
+                <div class="user-menu">
+                    <p class="user-name">Василий</p>
+                    <div class="popup-head">
+                        <ul class="popup-menu">
+                            <li class="menu-item">
+                                <a href="#" class="link">Настройки</a>
+                            </li>
+                            <li class="menu-item">
+                                <a href="#" class="link">Связаться с нами</a>
+                            </li>
+                            <li class="menu-item">
+                                <a href="#" class="link">Выход из системы</a>
+                            </li>
+
+                        </ul>
+                    </div>
+                </div>
             </div>
-        </div>
+        <?php endif; ?>
     </header>
 
     <main class="main-content container">
